Reformatted Course List added Date and Time. Moved things Around.
